move Sample::get_library tests from test_library to test_sample, delete empty test_library
